mini_projects/keil.gyozo/AI Meeting Assistant/meetingai/requirements.txt
# Core Framework
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.0.5

# LLM APIs
openai>=1.10.0
anthropic>=0.8.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Document Processing
python-docx>=1.1.0
python-markdown>=3.5.0

# Jira Integration
jira>=3.5.0

# Calendar Integration
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
msal>=1.25.0  # For Microsoft Graph

# Email
sendgrid>=6.11.0
python-dotenv>=1.0.0

# Utilities
pyyaml>=6.0.1
requests>=2.31.0
python-dateutil>=2.8.2

# Logging & Monitoring
colorlog>=6.8.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.12.0
flake8>=7.0.0
mypy>=1.7.0